added code tolimit length of ID user can input on the UpdateItem fragment
Gave the item ID input on the updateFragment a visual counter, then added code to that allows the app to reject inouts unless they are exactly 4 numerical characters long.
Also, all of the titles in the different ViewModel.java files has been changed from "this is the ___ fragment", to a more fitting title.